+++ packages/liveManager/src/recorder/IRecorder.ts
@@ -1,5 +1,6 @@
 import { EventEmitter } from "node:events";
 
+import type { VideoFormat } from "../index.js";
 import type { FormatName } from "./index.js";
 import type { XmlStreamController } from "../xml_stream_controller.js";
 
@@ -17,7 +18,7 @@ export interface BaseRecorderOptions {
   headers?: {
     [key: string]: string | undefined;
   };
-  videoFormat?: "auto" | "ts" | "mkv" | "mp4";
+  videoFormat?: VideoFormat;
 }
 
 /**
@@ -44,7 +45,12 @@ export interface IRecorder extends EventEmitter {
   // 事件类型定义
   on(
     event: "videoFileCreated",
-    listener: (data: { filename: string; cover?: string; rawFilename?: string }) => void,
+    listener: (data: {
+      filename: string;
+      cover?: string;
+      rawFilename?: string;
+      title?: string;
+    }) => void,
   ): this;
   on(event: "videoFileCompleted", listener: (data: { filename: string }) => void): this;
   on(event: "DebugLog", listener: (data: { type: string; text: string }) => void): this;
@@ -53,7 +59,7 @@ export interface IRecorder extends EventEmitter {
 
   emit(
     event: "videoFileCreated",
-    data: { filename: string; cover?: string; rawFilename?: string },
+    data: { filename: string; cover?: string; rawFilename?: string; title?: string },
   ): boolean;
   emit(event: "videoFileCompleted", data: { filename: string }): boolean;
   emit(event: "DebugLog", data: { type: string; text: string }): boolean;