+++ packages/BilibiliRecorder/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Next
 
+- 支持 `15000` 画质
+
+# 1.9.0
+
 - `recordHandle` 新增参数 `recorderType`
+- 优化ffmpeg默认参数，fmp4使用m4s后缀 [#224](https://github.com/renmu123/biliLive-tools/pull/224)
+- 支持 `25000` 原画真彩画质
 
 # 1.8.0