+++ packages/DouYinRecorder/CHANGELOG.md
@@ -1,6 +1,10 @@
-# Next
+# 1.9.0
 
 - `recordHandle` 新增参数 `recorderType`
+- 修复礼物弹幕缺失的情况
+- 修复两种接口的返回类型值错误
+- 修复使用 random 接口时，获取流可能失败的情况
+- 录制：优化ffmpeg默认参数，fmp4使用m4s后缀 [#224](https://github.com/renmu123/biliLive-tools/pull/224)
 
 # 1.8.0